+++ Cargo.toml
@@ -34,10 +34,11 @@ strip = true
 [dependencies]
 deltachat_derive = { path = "./deltachat_derive" }
 deltachat-time = { path = "./deltachat-time" }
+deltachat-contact-utils = { path = "./deltachat-contact-utils" }
 format-flowed = { path = "./format-flowed" }
 ratelimit = { path = "./deltachat-ratelimit" }
 
-anyhow = "1"
+anyhow = { workspace = true }
 async-channel = "2.0.0"
 async-imap = { version = "0.9.7", default-features = false, features = ["runtime-tokio"] }
 async-native-tls = { version = "0.5", default-features = false, features = ["runtime-tokio"] }
@@ -67,7 +68,7 @@ mime = "0.3.17"
 num_cpus = "1.16"
 num-derive = "0.4"
 num-traits = "0.2"
-once_cell = "1.18.0"
+once_cell = { workspace = true }
 percent-encoding = "2.3"
 parking_lot = "0.12"
 pgp = { version = "0.11", default-features = false }
@@ -77,9 +78,9 @@ qrcodegen = "1.7.0"
 quick-xml = "0.31"
 quoted_printable = "0.5"
 rand = "0.8"
-regex = "1.10"
+regex = { workspace = true }
 reqwest = { version = "0.12.2", features = ["json"] }
-rusqlite = { version = "0.31", features = ["sqlcipher"] }
+rusqlite = { workspace = true, features = ["sqlcipher"] }
 rust-hsluv = "0.1"
 sanitize-filename = "0.5"
 serde_json = "1"
@@ -132,6 +133,7 @@ members = [
   "deltachat-repl",
   "deltachat-time",
   "format-flowed",
+  "deltachat-contact-utils",
 ]
 
 [[bench]]
@@ -161,6 +163,12 @@ harness = false
 [[bench]]
 name = "send_events"
 harness = false
+
+[workspace.dependencies]
+anyhow = "1"
+once_cell = "1.18.0"
+regex = "1.10"
+rusqlite = { version = "0.31" }
 
 [features]
 default = ["vendored"]
