+++ deltachat-contact-utils/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "deltachat-contact-utils"
+version = "0.1.0"
+edition = "2021"
+description = "Contact-related tools, like parsing vcards and sanitizing name and address"
+license = "MPL-2.0"
+# TODO maybe it should be called "deltachat-text-utils" or similar?
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = { workspace = true }
+once_cell = { workspace = true }
+regex = { workspace = true }
+rusqlite = { workspace = true } # Needed in order to `impl rusqlite::types::ToSql for EmailAddress`. Could easily be put behind a feature.
+
+[dev-dependencies]
+anyhow = { workspace = true, features = ["backtrace"] } # Enable `backtrace` feature in tests.
